# backend/public/openapi.yaml
openapi: 3.0.3
info:
  title: API Helpdesk (Laravel + MySQL)
  version: 1.0.0
  description: 'API REST de gestion de tickets d''assistance : authentification par jetons Sanctum, CRUD, filtres, tri, pagination,
    relations imbriquées, fichiers, corbeille, statistiques.


    Comptes de démonstration (mot de passe `password`) : admin@demo.test, agent@demo.test, user@demo.test.


    Format d''erreur uniforme : `{"message": "...", "code": "identifiant_stable", ...}` ; en 422 s''ajoute `errors` (par champ).'
servers:
- url: http://127.0.0.1:8000/api/v1
  description: Serveur local (php artisan serve)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Jeton obtenu par /auth/connexion : Authorization: Bearer <jeton>'
  schemas:
    Erreur:
      type: object
      required:
      - message
      - code
      properties:
        message:
          type: string
          example: Transition impossible.
        code:
          type: string
          example: transition_invalide
        errors:
          type: object
          description: 'Présent en 422 : tableau de messages par champ'
          additionalProperties:
            type: array
            items:
              type: string
paths:
...
  /pieces-jointes/{id}:
    get:
      summary: Télécharger un fichier
      tags:
      - Pièces jointes
      description: 'Accès : **connecté**.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Succès
        '401':
          description: Non authentifié (jeton absent, invalide ou expiré)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erreur'
        '403':
          description: Interdit (droits ou capacité du jeton)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erreur'
        '404':
          description: Introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erreur'
      security:
      - bearerAuth: []
    delete:
      summary: Supprimer un fichier (déposant ou admin)
      tags:
      - Pièces jointes
      description: 'Accès : **connecté**.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: Succès sans contenu
        '401':
          description: Non authentifié (jeton absent, invalide ou expiré)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erreur'
        '403':
          description: Interdit (droits ou capacité du jeton)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erreur'
        '404':
          description: Introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erreur'
      security:
      - bearerAuth: []
...